requirements.txt
aws-cdk-lib>=2.244.0,<3.0.0
constructs>=10.5.0,<11.0.0
boto3>=1.26.0
pandas>=1.5.0
scikit-learn>=1.0.0
joblib>=1.3.2
